add a test case to test counter cache
