+++ test/lib/models/membership.rb
@@ -1,0 +1,3 @@
+class Membership < ActiveRecord::Base
+  belongs_to :organization, counter_cache: :memberships_count
+end
