+++ test/lib/models/organization.rb
@@ -1,0 +1,3 @@
+class Organization < ActiveRecord::Base
+  has_many :memberships, dependent: :destroy
+end
