+++ test/lib/models/post.rb
@@ -1,0 +1,4 @@
+class Post < ActiveRecord::Base
+  belongs_to :user
+  has_many :post_comments
+end
