+++ test/lib/models/post_comment.rb
@@ -1,0 +1,3 @@
+class PostComment < ActiveRecord::Base
+  belongs_to :post
+end
