+++ test/lib/sqlite3_connection.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+ActiveRecord::Base.establish_connection(
+  "adapter"  => "sqlite3",
+  "database" => ":memory:"
+)
